Alternative class construction
This removes `createClass` and `spec` as well as the dependency on
`create-react-class` in favor of `component` and `pureComponent` which
creates constructors that inherit from `React.Component` and
`React.PureComponent` in an idiomatic way for React. This also obviates
the need for some machinery like `ref` handling, which can be
implemented with `purescript-refs` instead.